+++ MinecraftServerLogsParsing/MinecraftServerLogsMapReduceJavaScript/src/main/java/org/timattt/jobs/onlyThread/OnlyThreadMapper.java
@@ -1,15 +1,14 @@
 package org.timattt.jobs.onlyThread;
 
-import org.apache.hadoop.io.DoubleWritable;
 import org.apache.hadoop.io.IntWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Mapper;
 
 import java.io.IOException;
 
-public class OnlyThreadMapper extends Mapper<Object, Text, Text, DoubleWritable> {
+public class OnlyThreadMapper extends Mapper<Object, Text, Text, IntWritable> {
     private final Text word = new Text();
-    private final DoubleWritable intWritable = new DoubleWritable();
+    private final IntWritable intWritable = new IntWritable();
 
     public void map(Object key, Text value, Context context) throws IOException, InterruptedException {
         System.out.println("INPUT ----------------> [" + key + "], [" + value.toString() + "]");
